Move create_grid
